Move ErrorTypePath into its own module
